refactor: colorIndex 공통이라 types 파일로 분리  및 적용
