+++ frontend/src/features/mindmap/node/types/node.ts
@@ -1,0 +1,4 @@
+export type NodeComponentProps = {
+    colorIndex?: number;
+    className?: string;
+};
